EZP-24361: Extracted RichText rendering to dedicated Twig extension
